// japfa-payload/payload-types.ts
export interface User {
  id: number;
  role: 'ADMIN' | 'OPERATION' | 'FARM';
  fullName: string;
  phone?: string | null;
  accountStatus: 'ACTIVE' | 'LOCKED';
  tenantMemberships?: (number | Tenant)[] | null;
  tenants?: (number | Tenant)[] | null;
  authzVersion: number;
  mustChangePassword: boolean;
  primaryTenant?: (number | null) | Tenant;
  createdBy?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tenants".
 */
export interface Tenant {
  id: number;
  /**
   * VD: FARM-BD-01, FAC-LA-02
   */
  farmCode: string;
  tenantId: string;
  name: string;
  system?: ('WMS' | 'FMS' | 'OMS' | 'TMS') | null;
  type: 'FACTORY' | 'FARM';
  status: 'ACTIVE' | 'INACTIVE' | 'LOCKED';
  flockId?: string | null;
  updatedBy?: string | null;
  managedBy?: string | null;
  managedByUser?: (number | null) | User;
  addressOfTenant?: string | null;
  phone?: string | null;
  flockName?: string | null;
  standardsApplied?: string | null;
  startFlockCount?: number | null;
  startFlockDate?: string | null;
  updatedAt: string;
  createdAt: string;
}